- removed Blazor.Diagrams, replaced with Neuroglia.Blazor.Dagre
