--- src/dashboard/Synapse.Dashboard/Features/Shared/WorkflowDiagram/EndNode.razor
@@ -1,18 +1,0 @@
-﻿@using Blazor.Diagrams.Components.Renderers
-@namespace Synapse.Dashboard
-
-<div class=end-node-container>
-    <div class="border border-5 border-dark rounded-circle bg-white end-node">
-        @foreach (var port in Node.Ports)
-        {
-            <PortRenderer Port="port" />
-        }
-    </div>
-</div>
-
-@code {
-
-    [Parameter]
-    public EndNodeModel Node { get; set; }
-
-}
