--- src/dashboard/Synapse.Dashboard/Features/Shared/WorkflowDiagram/EndNode.razor.css
@@ -1,9 +1,0 @@
-﻿.end-node-container {
-  width: 200px;
-}
-
-.end-node {
-  width: 35px;
-  height: 35px;
-  margin: auto;
-}
