--- src/dashboard/Synapse.Dashboard/Features/Shared/WorkflowDiagram/EndNodeModel.cs
@@ -1,23 +1,0 @@
-﻿using Blazor.Diagrams.Core.Models;
-using ServerlessWorkflow.Sdk.Models;
-
-namespace Synapse.Dashboard
-{
-    /// <summary>
-    /// Represents a <see cref="EndDefinition"/> <see cref="NodeModel"/>
-    /// </summary>
-    public class EndNodeModel
-        : WorkflowNodeModel
-    {
-
-        /// <summary>
-        /// Initializes a new <see cref="EndNodeModel"/>
-        /// </summary>
-        public EndNodeModel()
-        {
-            this.AddPort(PortAlignment.Top);
-        }
-
-    }
-
-}
