+++ src/dashboard/Synapse.Dashboard/Features/Shared/WorkflowDiagram/EndNodeViewModel.cs
@@ -4,7 +4,7 @@ using Neuroglia.Blazor.Dagre.Models;
 namespace Synapse.Dashboard
 {
     public class EndNodeViewModel
-        : NodeViewModel
+        : WorkflowNodeViewModel
     {
         public EndNodeViewModel()
             :base("", "end-node", NodeShape.Circle, 20, 20)
