--- src/dashboard/Synapse.Dashboard/Features/Shared/WorkflowDiagram/FunctionRefNode.razor.css
@@ -1,11 +1,0 @@
-﻿.function-node {
-  width: 200px;
-  font-size: 0.9rem;
-}
-
-.instance-counter {
-  width: 30px;
-  height: 30px;
-  bottom: -10px;
-  right: -10px;
-}
