+++ src/dashboard/Synapse.Dashboard/Features/Shared/WorkflowDiagram/FunctionRefNodeViewModel.cs
@@ -15,7 +15,7 @@ namespace Synapse.Dashboard
         /// <param name="action">The <see cref="ActionDefinition"/> the <see cref="FunctionRefNodeViewModel"/> represents</param>
         /// <param name="function">The <see cref="FunctionReference"/> the <see cref="FunctionRefNodeViewModel"/> represents</param>
         public FunctionRefNodeViewModel(ActionDefinition action, FunctionReference function) 
-            : base(action)
+            : base(action, "function-node")
         {
             this.Function = function;
         }
