+++ src/dashboard/Synapse.Dashboard/Features/Shared/WorkflowDiagram/JunctionNodeViewModel.cs
