--- src/dashboard/Synapse.Dashboard/Features/Shared/WorkflowDiagram/StartNode.razor
@@ -1,30 +1,0 @@
-﻿@using Blazor.Diagrams.Components.Renderers
-@namespace Synapse.Dashboard
-
-<div class=start-node-container>
-    <div class="border border-dark rounded-circle bg-white start-node">
-        @foreach (var port in Node.Ports)
-        {
-            <PortRenderer Port="port" />
-        }
-    </div>
-    <div class="border border-dark rounded-circle bg-primary text-white position-absolute instance-counter d-flex @DisplayClass"><span class="m-auto">@this.Node.ActiveInstances.Count</span></div>
-</div>
-
-@code {
-
-    [Parameter]
-    public StartNodeModel Node { get; set; }
-
-    string DisplayClass
-    {
-        get
-        {
-            if (this.Node.ActiveInstances.Any())
-                return string.Empty;
-            else
-                return "d-none";
-        }
-    }
-
-}
