--- src/dashboard/Synapse.Dashboard/Features/Shared/WorkflowDiagram/StartNode.razor.css
@@ -1,16 +1,0 @@
-﻿.start-node-container {
-  width: 200px;
-}
-
-.start-node {
-  width: 35px;
-  height: 35px;
-  margin: auto;
-}
-
-.instance-counter {
-  width: 30px;
-  height: 30px;
-  bottom: -10px;
-  right: -10px;
-}
