--- src/dashboard/Synapse.Dashboard/Features/Shared/WorkflowDiagram/StartNodeModel.cs
@@ -1,23 +1,0 @@
-﻿using Blazor.Diagrams.Core.Models;
-using ServerlessWorkflow.Sdk.Models;
-
-namespace Synapse.Dashboard
-{
-    /// <summary>
-    /// Represents a <see cref="StartDefinition"/> <see cref="NodeModel"/>
-    /// </summary>
-    public class StartNodeModel
-        : WorkflowNodeModel
-    {
-
-        /// <summary>
-        /// Initializes a new <see cref="WorkflowStateNodeModel"/>
-        /// </summary>
-        public StartNodeModel()
-        {
-            this.AddPort(PortAlignment.Bottom);
-        }
-
-    }
-
-}
