+++ src/dashboard/Synapse.Dashboard/Program.cs
@@ -49,7 +49,6 @@ builder.Services.AddSingleton<IAccordionManager, AccordionManager>();
 builder.Services.AddSingleton<IIntegrationEventStream, IntegrationEventStream>();
 builder.Services.AddSingleton<IMonacoEditorHelper, MonacoEditorHelper>();
 builder.Services.AddSingleton<IBreadcrumbService, BreadcrumbService>();
-builder.Services.AddScoped<IWorkflowDiagramBuilder, WorkflowDiagramBuilder>();
 builder.Services.AddScoped<IDagreService, DagreService>();
 builder.Services.AddScoped<WorkflowGraphBuilder>();
 builder.Services.AddFlux(flux =>
