+++ src/dashboard/Synapse.Dashboard/Synapse.Dashboard.csproj
@@ -39,7 +39,6 @@
   <ItemGroup>
 	<PackageReference Include="BlazorMonaco" Version="2.1.0" />
 	<PackageReference Include="BuildWebCompiler2022" Version="1.14.7" />
-	<PackageReference Include="GraphShape" Version="1.2.1" />
     <PackageReference Include="Microsoft.AspNetCore.Components.WebAssembly" Version="6.0.3" />
     <PackageReference Include="Microsoft.AspNetCore.Components.WebAssembly.DevServer" Version="6.0.3" PrivateAssets="all" />
     <PackageReference Include="Microsoft.AspNetCore.SignalR.Client" Version="6.0.3" />
@@ -48,7 +47,6 @@
 	<PackageReference Include="Neuroglia.Data.Flux" Version="2.0.1.57" />
 	<PackageReference Include="Neuroglia.Data.Flux.ReduxDevTools" Version="2.0.1.57" />
 	<PackageReference Include="Neuroglia.Data.OData" Version="2.0.1.57" />
-	<PackageReference Include="Z.Blazor.Diagrams" Version="2.1.6" />
   </ItemGroup>
 
   <ItemGroup>
