+++ src/dashboard/Synapse.Dashboard/_Imports.razor
@@ -18,5 +18,4 @@
 @using Synapse.Dashboard.Services
 @using Synapse.Integration.Models
 
-
 @using BlazorMonaco
