+++ src/dashboard/Synapse.Dashboard/wwwroot/css/app.css
@@ -12282,3 +12282,7 @@ input[type=time]::-webkit-clear-button {
 .oi {
   width: 16px;
 }
+
+.end-node circle {
+  stroke-width: 3px;
+}
