+++ src/dashboard/Synapse.Dashboard/wwwroot/css/app.scss
@@ -72,3 +72,7 @@ input[type=time]::-webkit-clear-button {
 .oi {
     width: 16px;
 }
+
+.end-node circle {
+    stroke-width: 3px;
+}
