+++ src/dashboard/Synapse.Dashboard/wwwroot/index.html
@@ -6,12 +6,10 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0, maximum-scale=1.0, user-scalable=no" />
     <title>Synapse</title>
     <base href="/" />
-    <link href="css/app.css" rel="stylesheet" />
-    <link href="Synapse.Dashboard.styles.css" rel="stylesheet" />
-    <link href="_content/Z.Blazor.Diagrams/style.min.css" rel="stylesheet" />
-    <link href="_content/Z.Blazor.Diagrams/default.styles.min.css" rel="stylesheet" />
     <link href="_content/Neuroglia.Blazor.Dagre/neuroglia-blazor-dagre.min.css" rel="stylesheet" />
     <link href="_content/BlazorMonaco/lib/monaco-editor/min/vs/editor/editor.main.css" rel="stylesheet" />
+    <link href="Synapse.Dashboard.styles.css" rel="stylesheet" />
+    <link href="css/app.css" rel="stylesheet" />
     <link href="manifest.json" rel="manifest" />
     <link rel="apple-touch-icon" sizes="512x512" href="icon-512.png" />
     <link rel="apple-touch-icon" sizes="192x192" href="icon-192.png" />
@@ -54,14 +52,14 @@
     <div id="app" class="h-100">
         <div id="logo-container">
             <svg class="logo-viewbox" viewBox="0 0 612 567" xml:space="preserve">
-            <use xlink:href="#brain" class="orange animation-phase-1" />
-            <use xlink:href="#s1-definition" class="blue animation-phase-2" />
-            <use xlink:href="#y-definition" class="blue animation-phase-3" />
-            <use xlink:href="#n-definition" class="blue animation-phase-4" />
-            <use xlink:href="#a-definition" class="blue animation-phase-5" />
-            <use xlink:href="#p-definition" class="blue animation-phase-6" />
-            <use xlink:href="#s2-definition" class="blue animation-phase-7" />
-            <use xlink:href="#e-definition" class="blue animation-phase-8" />
+                <use xlink:href="#brain" class="orange animation-phase-1" />
+                <use xlink:href="#s1-definition" class="blue animation-phase-2" />
+                <use xlink:href="#y-definition" class="blue animation-phase-3" />
+                <use xlink:href="#n-definition" class="blue animation-phase-4" />
+                <use xlink:href="#a-definition" class="blue animation-phase-5" />
+                <use xlink:href="#p-definition" class="blue animation-phase-6" />
+                <use xlink:href="#s2-definition" class="blue animation-phase-7" />
+                <use xlink:href="#e-definition" class="blue animation-phase-8" />
             </svg>
             <div class="spinner-border" role="status">
                 <span class="visually-hidden">Loading...</span>
@@ -71,14 +69,14 @@
 
     <div id="animation-container" class="fade-out">
         <svg id="splash-viewbox" class="logo-viewbox" viewBox="0 0 612 567" xml:space="preserve">
-        <use xlink:href="#brain" class="orange animation-phase-1" />
-        <use xlink:href="#s1-definition" class="blue animation-phase-2" />
-        <use xlink:href="#y-definition" class="blue animation-phase-3" />
-        <use xlink:href="#n-definition" class="blue animation-phase-4" />
-        <use xlink:href="#a-definition" class="blue animation-phase-5" />
-        <use xlink:href="#p-definition" class="blue animation-phase-6" />
-        <use xlink:href="#s2-definition" class="blue animation-phase-7" />
-        <use xlink:href="#e-definition" class="blue animation-phase-8" />
+            <use xlink:href="#brain" class="orange animation-phase-1" />
+            <use xlink:href="#s1-definition" class="blue animation-phase-2" />
+            <use xlink:href="#y-definition" class="blue animation-phase-3" />
+            <use xlink:href="#n-definition" class="blue animation-phase-4" />
+            <use xlink:href="#a-definition" class="blue animation-phase-5" />
+            <use xlink:href="#p-definition" class="blue animation-phase-6" />
+            <use xlink:href="#s2-definition" class="blue animation-phase-7" />
+            <use xlink:href="#e-definition" class="blue animation-phase-8" />
         </svg>
     </div>
 
@@ -90,7 +88,6 @@
     <script src="js/splash.js"></script>
     <script src="_framework/blazor.webassembly.js"></script>
     <script>navigator.serviceWorker.register('service-worker.js');</script>
-    <script src="_content/Z.Blazor.Diagrams/script.min.js"></script>
     <script src="_content/Neuroglia.Blazor.Dagre/lib/dagre/dagre.js"></script>
     <script src="_content/Neuroglia.Blazor.Dagre/neuroglia-blazor-dagre-interop.js"></script>
     <script src="js/monaco-editor-extensions.js"></script>
